split ZMQ wire module
