+++ lib/kv-router/src/zmq_wire/tests.rs
@@ -1,0 +1,377 @@
+// SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use std::sync::Arc;
+use std::sync::atomic::AtomicU32;
+
+use rmp_serde::{from_slice, to_vec};
+use rstest::rstest;
+
+use crate::protocols::{
+    BlockHashOptions, ExternalSequenceBlockHash, KvCacheEventData, WorkerWithDpRank,
+    compute_block_hash_for_seq,
+};
+
+use super::*;
+
+#[derive(Clone, Copy, Debug)]
+enum TestEventKind {
+    BlockStored,
+    BlockRemoved,
+}
+
+#[test]
+fn test_deserialize_bigram_block_stored_sequence() {
+    let raw_event = (
+        "BlockStored",
+        vec![BlockHashValue::Unsigned(11), BlockHashValue::Unsigned(12)],
+        Option::<BlockHashValue>::None,
+        vec![(10u32, 11u32), (11, 12), (12, 13), (13, 14)],
+        2usize,
+        Option::<u64>::None,
+        Option::<String>::None,
+        Option::<String>::None,
+    );
+    let encoded = to_vec(&raw_event).unwrap();
+    let event: RawKvEvent = from_slice(&encoded).unwrap();
+
+    match event {
+        RawKvEvent::BlockStored {
+            token_ids,
+            block_size,
+            is_eagle,
+            ..
+        } => {
+            assert_eq!(token_ids, vec![10, 11, 12, 13, 14]);
+            assert_eq!(block_size, 2);
+            assert_eq!(is_eagle, Some(true));
+        }
+        other => panic!("expected BlockStored, got {other:?}"),
+    }
+}
+
+fn block_stored_sequence(
+    group_idx: Option<u32>,
+    kv_cache_spec_kind: Option<&'static str>,
+) -> Vec<u8> {
+    match (group_idx, kv_cache_spec_kind) {
+        (Some(group_idx), Some(kv_cache_spec_kind)) => to_vec(&(
+            "BlockStored",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<BlockHashValue>::None,
+            vec![10u32, 11],
+            2usize,
+            Option::<u64>::None,
+            Option::<String>::None,
+            Option::<String>::None,
+            Option::<u8>::None,
+            group_idx,
+            kv_cache_spec_kind,
+        ))
+        .unwrap(),
+        (Some(group_idx), None) => to_vec(&(
+            "BlockStored",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<BlockHashValue>::None,
+            vec![10u32, 11],
+            2usize,
+            Option::<u64>::None,
+            Option::<String>::None,
+            Option::<String>::None,
+            Option::<u8>::None,
+            group_idx,
+        ))
+        .unwrap(),
+        (None, Some(kv_cache_spec_kind)) => to_vec(&(
+            "BlockStored",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<BlockHashValue>::None,
+            vec![10u32, 11],
+            2usize,
+            Option::<u64>::None,
+            Option::<String>::None,
+            Option::<String>::None,
+            Option::<u8>::None,
+            Option::<u32>::None,
+            kv_cache_spec_kind,
+        ))
+        .unwrap(),
+        (None, None) => to_vec(&(
+            "BlockStored",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<BlockHashValue>::None,
+            vec![10u32, 11],
+            2usize,
+            Option::<u64>::None,
+            Option::<String>::None,
+            Option::<String>::None,
+        ))
+        .unwrap(),
+    }
+}
+
+fn block_removed_sequence(
+    group_idx: Option<u32>,
+    kv_cache_spec_kind: Option<&'static str>,
+) -> Vec<u8> {
+    match (group_idx, kv_cache_spec_kind) {
+        (Some(group_idx), Some(kv_cache_spec_kind)) => to_vec(&(
+            "BlockRemoved",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<String>::None,
+            group_idx,
+            kv_cache_spec_kind,
+        ))
+        .unwrap(),
+        (Some(group_idx), None) => to_vec(&(
+            "BlockRemoved",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<String>::None,
+            group_idx,
+        ))
+        .unwrap(),
+        (None, Some(kv_cache_spec_kind)) => to_vec(&(
+            "BlockRemoved",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<String>::None,
+            Option::<u32>::None,
+            kv_cache_spec_kind,
+        ))
+        .unwrap(),
+        (None, None) => to_vec(&(
+            "BlockRemoved",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<String>::None,
+        ))
+        .unwrap(),
+    }
+}
+
+fn sequence_with_group_idx(event_kind: TestEventKind, group_idx: Option<u32>) -> Vec<u8> {
+    match event_kind {
+        TestEventKind::BlockStored => block_stored_sequence(group_idx, None),
+        TestEventKind::BlockRemoved => block_removed_sequence(group_idx, None),
+    }
+}
+
+fn sequence_with_cache_spec_kind(
+    event_kind: TestEventKind,
+    group_idx: Option<u32>,
+    kv_cache_spec_kind: &'static str,
+) -> Vec<u8> {
+    match event_kind {
+        TestEventKind::BlockStored => block_stored_sequence(group_idx, Some(kv_cache_spec_kind)),
+        TestEventKind::BlockRemoved => block_removed_sequence(group_idx, Some(kv_cache_spec_kind)),
+    }
+}
+
+fn sequence_with_cache_spec_kind_without_group_idx_slot(
+    event_kind: TestEventKind,
+    kv_cache_spec_kind: &'static str,
+) -> Vec<u8> {
+    match event_kind {
+        TestEventKind::BlockStored => to_vec(&(
+            "BlockStored",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<BlockHashValue>::None,
+            vec![10u32, 11],
+            2usize,
+            Option::<u64>::None,
+            Option::<String>::None,
+            Option::<String>::None,
+            Option::<u8>::None,
+            kv_cache_spec_kind,
+        ))
+        .unwrap(),
+        TestEventKind::BlockRemoved => to_vec(&(
+            "BlockRemoved",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<String>::None,
+            kv_cache_spec_kind,
+        ))
+        .unwrap(),
+    }
+}
+
+fn sequence_with_cache_spec_kind_and_sliding_window(
+    event_kind: TestEventKind,
+    group_idx: u32,
+    kv_cache_spec_kind: &'static str,
+    kv_cache_spec_sliding_window: u32,
+) -> Vec<u8> {
+    match event_kind {
+        TestEventKind::BlockStored => to_vec(&(
+            "BlockStored",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<BlockHashValue>::None,
+            vec![10u32, 11],
+            2usize,
+            Option::<u64>::None,
+            Option::<String>::None,
+            Option::<String>::None,
+            Option::<u8>::None,
+            group_idx,
+            kv_cache_spec_kind,
+            kv_cache_spec_sliding_window,
+        ))
+        .unwrap(),
+        TestEventKind::BlockRemoved => to_vec(&(
+            "BlockRemoved",
+            vec![BlockHashValue::Unsigned(11)],
+            Option::<String>::None,
+            group_idx,
+            kv_cache_spec_kind,
+            kv_cache_spec_sliding_window,
+        ))
+        .unwrap(),
+    }
+}
+
+fn assert_parsed_event_kind(event: RawKvEvent, expected_kind: TestEventKind) {
+    match (event, expected_kind) {
+        (RawKvEvent::BlockStored { .. }, TestEventKind::BlockStored)
+        | (RawKvEvent::BlockRemoved { .. }, TestEventKind::BlockRemoved) => {}
+        (event, expected_kind) => {
+            panic!("expected {expected_kind:?}, got {event:?}");
+        }
+    }
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_accepts_main_group_idx(#[case] event_kind: TestEventKind) {
+    let event: RawKvEvent = from_slice(&sequence_with_group_idx(event_kind, Some(0))).unwrap();
+
+    assert_parsed_event_kind(event, event_kind);
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_ignores_non_main_group_idx(#[case] event_kind: TestEventKind) {
+    let event: RawKvEvent = from_slice(&sequence_with_group_idx(event_kind, Some(1))).unwrap();
+
+    assert!(matches!(event, RawKvEvent::Ignored));
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_accepts_missing_group_idx(#[case] event_kind: TestEventKind) {
+    let event: RawKvEvent = from_slice(&sequence_with_group_idx(event_kind, None)).unwrap();
+
+    assert_parsed_event_kind(event, event_kind);
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_accepts_main_attention_kind_with_nonzero_group_idx(
+    #[case] event_kind: TestEventKind,
+) {
+    let event: RawKvEvent = from_slice(&sequence_with_cache_spec_kind(
+        event_kind,
+        Some(3),
+        "full_attention",
+    ))
+    .unwrap();
+
+    assert_parsed_event_kind(event, event_kind);
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_accepts_main_attention_kind_without_group_idx_slot(
+    #[case] event_kind: TestEventKind,
+) {
+    let event: RawKvEvent = from_slice(&sequence_with_cache_spec_kind_without_group_idx_slot(
+        event_kind,
+        "full_attention",
+    ))
+    .unwrap();
+
+    assert_parsed_event_kind(event, event_kind);
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_accepts_main_attention_kind_with_sliding_window(
+    #[case] event_kind: TestEventKind,
+) {
+    let event: RawKvEvent = from_slice(&sequence_with_cache_spec_kind_and_sliding_window(
+        event_kind,
+        3,
+        "full_attention",
+        128,
+    ))
+    .unwrap();
+
+    assert_parsed_event_kind(event, event_kind);
+}
+
+#[rstest]
+#[case(TestEventKind::BlockStored)]
+#[case(TestEventKind::BlockRemoved)]
+fn test_deserialize_sequence_ignores_non_main_attention_kind_with_group_idx_zero(
+    #[case] event_kind: TestEventKind,
+) {
+    let event: RawKvEvent =
+        from_slice(&sequence_with_cache_spec_kind(event_kind, Some(0), "mamba")).unwrap();
+
+    assert!(matches!(event, RawKvEvent::Ignored));
+}
+
+#[test]
+fn test_convert_event_bigram_emits_eagle_windows() {
+    let raw_event = RawKvEvent::BlockStored {
+        block_hashes: vec![BlockHashValue::Unsigned(21), BlockHashValue::Unsigned(22)],
+        parent_block_hash: None,
+        token_ids: vec![10, 11, 12, 13, 14],
+        block_size: 2,
+        medium: None,
+        lora_name: None,
+        block_mm_infos: None,
+        is_eagle: Some(true),
+    };
+    let warning_count = Arc::new(AtomicU32::new(0));
+    let placement_event =
+        convert_event(raw_event, 7, 2, WorkerWithDpRank::new(3, 0), &warning_count);
+
+    match placement_event.unwrap().event.data {
+        KvCacheEventData::Stored(store_data) => {
+            assert_eq!(store_data.blocks.len(), 2);
+            assert_eq!(
+                store_data.blocks[0].block_hash,
+                ExternalSequenceBlockHash(21)
+            );
+            assert_eq!(
+                store_data.blocks[1].block_hash,
+                ExternalSequenceBlockHash(22)
+            );
+
+            let expected_first = compute_block_hash_for_seq(
+                &[10, 11, 12],
+                2,
+                BlockHashOptions {
+                    is_eagle: Some(true),
+                    ..Default::default()
+                },
+            );
+            let expected_second = compute_block_hash_for_seq(
+                &[12, 13, 14],
+                2,
+                BlockHashOptions {
+                    is_eagle: Some(true),
+                    ..Default::default()
+                },
+            );
+
+            assert_eq!(store_data.blocks[0].tokens_hash, expected_first[0]);
+            assert_eq!(store_data.blocks[1].tokens_hash, expected_second[0]);
+        }
+        other => panic!("expected Stored event, got {other:?}"),
+    }
+}
